icons: Ensure symlinks are not pointing to other symlinks
While this works, it may lead to troubles when a synlink is supposed to
be created by upcoming definitions, so let's avoid this

+++ icons/src/symlinks/generate-symlinks.sh
@@ -144,6 +144,10 @@ linker () {
 				echo "  ERROR: Can't link a file with itself!"
 				exit 1
 			fi
+			if [ -L "$target" ]; then
+				echo "  ERROR: $target is already a symlink, please point it to a real file!"
+				exit 1
+			fi
 			if [ -z "$_dry_run" ]; then
 				ln -sf "$target" "$link_name"
 			fi
